Refactors icons and links components file locations
