+++ src/app/global/icons/style/icons.scss
