+++ src/app/top/Top.jsx
@@ -2,9 +2,9 @@ import React from 'react'
 import { Route, Redirect, Switch } from 'react-router-dom'
 import { Footer } from './Footer'
 import { NavBar } from './Navbar'
-import { About } from '../about/About'
-import { Portfolio } from '../portfolio/Portfolio'
-import { EmailForm } from '../contact/EmailForm'
+import { About } from './about/About'
+import { Portfolio } from './portfolio/Portfolio'
+import { EmailForm } from './contact/EmailForm'
 import './style/global'
 
 const routesParams = {
@@ -22,7 +22,7 @@ const routesParams = {
   },
 }
 
-export const App = () => {
+export const Top = () => {
   // Routes object
   const routes = Object.keys(routesParams).map(key => {
     return (
