+++ src/app/top/about/education/style/education.scss
