+++ src/app/top/portfolio/applications/Applications.jsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { Application } from './Application'
-import appsData from '../../../global/data/applications'
-import externalLinksData from '../../../global/data/external-links'
+import appsData from '../../../../global/data/applications'
+import externalLinksData from '../../../../global/data/external-links'
 
 export const Applications = () => {
   // Apps data object
