+++ src/app/top/portfolio/blog/BlogPostPreview.jsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import { useRouteMatch } from 'react-router-dom'
-import { NavLinkTop } from '../../../global/components/link/NavLinkTop'
+import { NavLinkTop } from '../../../global/links/NavLinkTop'
 import './style/blog-post-preview'
 
 export const BlogPostPreview = ({ blogPostData, blogPostDataPath, postId }) => {
