+++ src/app/top/portfolio/blog/style/blog-post-preview.scss
