+++ src/app/top/portfolio/featured/FeaturedContentCard.jsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import { WebLink } from '../../../global/components/link/WebLink'
+import { WebLink } from '../../../global/links/WebLink'
 
 export const FeaturedContentCard = ({ featuredContentItem }) => {
   const card = (
